+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog
 
 ## [Unreleased]
+
 - Added secutiry
+- Fix error message string interpolation
 
 ## v.0.3.0 - 2020-06-09
+
 - Improved icon
 - Add OTA task for active file